[diffusion, rollout] refactor: move request-batch helpers out of utils
Keep pipelines/utils.py for training dispatch; collate/split helpers for
request-level diffusion batching now live in pipelines/request_batch.py.
Also add the enablement doc already linked from FlowGRPO / index.
